requirements.txt
numpy
opencv-python
torch
torchvision
matplotlib
tqdm
pillow
